Refactor: Pkgdata relation module
